pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tiny-transformer-build"
version = "0.1.0"
description = "Educational repository for building transformer models from scratch"
readme = "README.md"
authors = [
    {name = "Tiny Transformer Build Contributors"}
]
license = {text = "MIT"}
requires-python = ">=3.9"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Education",
]
keywords = ["transformers", "attention", "deep-learning", "pytorch", "education", "tutorial"]

dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "matplotlib>=3.7.0",
    "tqdm>=4.66.0",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pandas>=2.0.0",
    "pyyaml>=6.0.0",
    "typing-extensions>=4.8.0",
]

[project.optional-dependencies]
dev = [
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "notebook>=7.0.0",
    "ipykernel>=6.25.0",
    "nbclient>=0.8.0",
    "nbformat>=5.9.0",
]
viz = [
    "scikit-learn>=1.0.0",  # For dimensionality reduction in visualizations
]
all = [
    "tiny-transformer-build[dev,notebooks,viz]",
]

[project.urls]
Homepage = "https://github.com/yourusername/tiny-transformer-build"
Documentation = "https://github.com/yourusername/tiny-transformer-build#readme"
Repository = "https://github.com/yourusername/tiny-transformer-build"
Issues = "https://github.com/yourusername/tiny-transformer-build/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["tiny_transformer*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "gpu: marks tests requiring GPU",
]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

[tool.coverage.run]
source = ["tiny_transformer"]
omit = ["tests/*", ".venv/*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
